+++ Sources/MLXAudioCore/MLXAudio.swift
@@ -3,7 +3,7 @@
 /// MLXAudio library version and metadata
 public enum MLXAudio {
     /// Current version of the MLXAudio library
-    public static let version = "0.8.7"
+    public static let version = "0.9.0"
 
     /// Library name
     public static let name = "MLXAudio"